requirements.txt
python-dotenv
langchain-text-splitters
chromadb
google-genai